+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Fix $filter for navigation elements 
+- Fix \$filter for navigation elements
+- Add request authorization parsing for logging 
 
 ## Version 1.4.25 - 2020-04-08